+++ source/MRMesh/MRMesh.cpp
@@ -161,224 +161,6 @@ MeshEdgePoint Mesh::toEdgePoint( VertId v ) const
     return MeshEdgePoint( topology, v );
 }
 
-namespace
-{
-
-/// computes the summed six-fold volume of tetrahedrons with one vertex at (0,0,0) and other three vertices taken from a mesh's triangle
-class FaceVolumeCalc
-{
-public:
-    FaceVolumeCalc( const Mesh& mesh, const FaceBitSet& region) : mesh_( mesh ), region_( region )
-    {}
-    FaceVolumeCalc( FaceVolumeCalc& x, tbb::split ) : mesh_( x.mesh_ ), region_( x.region_ )
-    {}
-    void join( const FaceVolumeCalc& y )
-    {
-        volume_ += y.volume_;
-    }
-
-    double volume() const
-    {
-        return volume_;
-    }
-
-    void operator()( const tbb::blocked_range<FaceId>& r )
-    {
-        for ( FaceId f = r.begin(); f < r.end(); ++f )
-        {
-            if ( region_.test( f ) && mesh_.topology.hasFace( f ) )
-            {
-                const auto coords = mesh_.getTriPoints( f );
-                volume_ += mixed( Vector3d( coords[0] ), Vector3d( coords[1] ), Vector3d( coords[2] ) );
-            }
-        }
-    }
-
-private:
-    const Mesh& mesh_;
-    const FaceBitSet& region_;
-    double volume_{ 0.0 };
-};
-
-/// computes the summed six-fold volume of tetrahedrons with one vertex at (0,0,0), another vertex at the center of hole,
-/// and other two vertices taken from a hole's edge
-class HoleVolumeCalc
-{
-public:
-    HoleVolumeCalc( const Mesh& mesh, const std::vector<EdgeId>& holeRepresEdges ) : mesh_( mesh ), holeRepresEdges_( holeRepresEdges )
-    {}
-    HoleVolumeCalc( HoleVolumeCalc& x, tbb::split ) : mesh_( x.mesh_ ), holeRepresEdges_( x.holeRepresEdges_ )
-    {}
-    void join( const HoleVolumeCalc& y )
-    {
-        volume_ += y.volume_;
-    }
-
-    double volume() const
-    {
-        return volume_;
-    }
-
-    void operator()( const tbb::blocked_range<size_t>& r )
-    {
-        for ( size_t i = r.begin(); i < r.end(); ++i )
-        {
-            const auto e0 = holeRepresEdges_[i];
-            Vector3d sumBdPos;
-            int countBdVerts = 0;
-            for ( auto e : leftRing( mesh_.topology, e0 ) )
-            {
-                sumBdPos += Vector3d( mesh_.orgPnt( e ) );
-                ++countBdVerts;
-            }
-            Vector3d holeCenter = sumBdPos / double( countBdVerts );
-            for ( auto e : leftRing( mesh_.topology, e0 ) )
-            {
-                volume_ += mixed( holeCenter, Vector3d( mesh_.orgPnt( e ) ), Vector3d( mesh_.destPnt( e ) ) );
-            }
-        }
-    }
-
-private:
-    const Mesh& mesh_;
-    const std::vector<EdgeId>& holeRepresEdges_;
-    double volume_{ 0.0 };
-};
-
-} // anonymous namespace
-
-double Mesh::volume( const FaceBitSet* region /*= nullptr */ ) const
-{
-    MR_TIMER
-    const auto lastValidFace = topology.lastValidFace();
-    const auto& faces = topology.getFaceIds( region );
-    FaceVolumeCalc fcalc( *this, faces );
-    parallel_deterministic_reduce( tbb::blocked_range<FaceId>( 0_f, lastValidFace + 1, 1024 ), fcalc );
-
-    const auto holeRepresEdges = topology.findHoleRepresentiveEdges( region );
-    HoleVolumeCalc hcalc( *this, holeRepresEdges );
-    parallel_deterministic_reduce( tbb::blocked_range<size_t>( size_t( 0 ), holeRepresEdges.size() ), hcalc );
-
-    return ( fcalc.volume() + hcalc.volume() ) / 6.0;
-}
-
-double Mesh::holePerimiter( EdgeId e0 ) const
-{
-    double res = 0;
-    if ( topology.left( e0 ) )
-    {
-        assert( false );
-        return res;
-    }
-
-    for ( auto e : leftRing( topology, e0 ) )
-    {
-        assert( !topology.left( e ) );
-        res += edgeLength( e );
-    }
-    return res;
-}
-
-Vector3d Mesh::holeDirArea( EdgeId e0 ) const
-{
-    Vector3d sum;
-    if ( topology.left( e0 ) )
-    {
-        assert( false );
-        return sum;
-    }
-
-    Vector3d p0{ orgPnt( e0 ) };
-    for ( auto e : leftRing0( topology, e0 ) )
-    {
-        assert( !topology.left( e ) );
-        Vector3d p1{ orgPnt( e ) };
-        Vector3d p2{ destPnt( e ) };
-        sum += cross( p1 - p0, p2 - p0 );
-    }
-    return 0.5 * sum;
-}
-
-Vector3f Mesh::leftTangent( EdgeId e ) const
-{
-    assert( topology.left( e ) );
-    const auto lNorm = leftNormal( e );
-    const auto eDir = edgeVector( e ).normalized();
-    return cross( lNorm, eDir );
-}
-
-Vector3f Mesh::dirDblArea( VertId v ) const
-{
-    Vector3f sum;
-    for ( EdgeId e : orgRing( topology, v ) )
-    {
-        if ( topology.left( e ).valid() )
-        {
-            sum += leftDirDblArea( e );
-        }
-    }
-    return sum;
-}
-
-Vector3f Mesh::normal( const MeshTriPoint & p ) const
-{
-    VertId a, b, c;
-    topology.getLeftTriVerts( p.e, a, b, c );
-    auto n0 = normal( a );
-    auto n1 = normal( b );
-    auto n2 = normal( c );
-    return p.bary.interpolate( n0, n1, n2 ).normalized();
-}
-
-Vector3f Mesh::pseudonormal( VertId v, const FaceBitSet * region ) const
-{
-    Vector3f sum;
-    for ( EdgeId e : orgRing( topology, v ) )
-    {
-        const auto l = topology.left( e );
-        if ( l && ( !region || region->test( l ) ) )
-        {
-            auto d0 = edgeVector( e );
-            auto d1 = edgeVector( topology.next( e ) );
-            auto angle = MR::angle( d0, d1 );
-            auto n = cross( d0, d1 );
-            sum += angle * n.normalized();
-        }
-    }
-
-    return sum.normalized();
-}
-
-Vector3f Mesh::pseudonormal( UndirectedEdgeId ue, const FaceBitSet * region ) const
-{
-    EdgeId e{ ue };
-    auto l = topology.left( e );
-    if ( l && region && !region->test( l ) )
-        l = {};
-    auto r = topology.right( e );
-    if ( r && region && !region->test( r ) )
-        r = {};
-    if ( !l && !r )
-        return {};
-    if ( !l )
-        return normal( r );
-    if ( !r )
-        return normal( l );
-    auto nl = normal( l );
-    auto nr = normal( r );
-    return ( nl + nr ).normalized();
-}
-
-Vector3f Mesh::pseudonormal( const MeshTriPoint & p, const FaceBitSet * region ) const
-{
-    if ( auto v = p.inVertex( topology ) )
-        return pseudonormal( v, region );
-    if ( auto e = p.onEdge( topology ) )
-        return pseudonormal( e.e.undirected(), region );
-    assert( !region || region->test( topology.left( p.e ) ) );
-    return leftNormal( p.e );
-}
-
 bool Mesh::isOutsideByProjNorm( const Vector3f & pt, const MeshProjectionResult & proj, const FaceBitSet * region ) const
 {
     return dot( proj.proj.point - pt, pseudonormal( proj.mtp, region ) ) <= 0;
